+++ util/config_file.h
@@ -579,7 +579,7 @@ struct config_file {
 	/** ratelimit factor, 0 blocks all, 10 allows 1/10 of traffic */
 	int ratelimit_factor;
 	/** number of retries on outgoing queries */
-	size_t outbound_msg_retry;
+	int outbound_msg_retry;
 	/** minimise outgoing QNAME and hide original QTYPE if possible */
 	int qname_minimisation;
 	/** minimise QNAME in strict mode, minimise according to RFC.
